Add RayCast

+++ Assets/Scripts/SpaceShipEnemyFactory.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class SpaceShipEnemyFactory
 {
-    public SpaceShipEnemy Create(float speed, int damage, int health, float armor, Transform targetTransform)
+    public SpaceShipEnemy Create(float speed, int damage, int health, int armor, Transform targetTransform)
     {
         SpaceShipEnemy spaceShipEnemy = Object.Instantiate(Resources.Load<SpaceShipEnemy>("SpaceShipEnemy_0"));
         spaceShipEnemy.SpaceShipEnemyInitParametr(speed, damage, health, armor);
